Optimize creation process of generated payment slip for new orders
After the order is being made (and this plugin has been selected as payment option), WooCommerce is first sending order details to the user and after that is preparing and displaying so-called "thank you" page with order details. Since payment slip is being provided both in e-mail message as well as in thank you page, it has been until now generated for those both two steps. As those two steps are performed imediatelly one after another, there is no need to generate them two times - the better option is to store it first time temporarily in the webserver's session and to reuse it second time by fetching it from there.