Ensure that JavaScript functions are defined before they are used
when plugin was installed on several web-shops, on some of them sometimes the image wasn't properly displayed (empty parts of image weren't cropped) and the button for downloading payment slip image was not working. That was caused since image-cropping function definition was defined after the event handler that triggers it.