Add option to send e-mail messages without PHPMailer module
The reason for introducing this alternative way for sending e-mails was to resolve issue reported [here](https://wordpress.org/support/topic/kupci-ne-dobiju-mail-nakon-zavrsene-narudzbe/). I wasn't able to reproduce this problem by using FluentSMTP plugin, but in a topic discussion [here](https://wordpress.org/support/topic/woocommerce_email_attachments-content-id/) it was mentioned that when using dedicated mailing functionalities, there is a chance that some e-mail features might stop work. Therefore a new settings option was introduced which is by default enabled and then the PHPMailer module is used for sending e-mail. Otherwise the attachments for sending are stored in OS-specific temporary directory and are attached to e-mail message by specifying method for filter _woocommerce_email_attachments_ which is returning paths of temporary files to be attached. Drawback in second approach is that images cannot be display inline (i.e. as part of the message body), so they are treated and displayed as regular attachments.